requirements.txt
flask
flask_smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate    
cloud-sql-python-connector[pymysql]
pg8000
openai